v2.0 semi-working Base-class-triggered  builders not working yet Customizable templates not working yet

+++ src/Refresh.ps1
@@ -1,5 +1,10 @@
+Uninstall-Package -Id BuilderGenerator -ProjectName BuilderGenerator.Test.NuGet
+
+if (test-path C:\Users\Mel\.nuget\packages\buildergenerator\$version)
+{
+    rm C:\Users\Mel\.nuget\packages\buildergenerator\$version -recurse
+}
+
 ./Publish-Local.ps1
 
-Uninstall-Package -Id BuilderGenerator -ProjectName BuilderGenerator.Sample
-rm C:\Users\Mel\.nuget\packages\buildergenerator\$version -recurse
-Install-Package -Id BuilderGenerator -ProjectName BuilderGenerator.Sample
+Install-Package -Id BuilderGenerator -ProjectName BuilderGenerator.Test.NuGet
